+++ bundles/org.openhab.binding.somfymylink/src/main/java/org/openhab/binding/somfymylink/internal/model/SomfyMyLinkShade.java
@@ -10,7 +10,9 @@
  *
  * SPDX-License-Identifier: EPL-2.0
  */
-package org.openhab.binding.somfymylink.internal;
+package org.openhab.binding.somfymylink.internal.model;
+
+import org.openhab.binding.somfymylink.internal.SomfyMyLinkBindingConstants;
 
 /**
  * The {@link SomfyMyLinkBindingConstants} class defines common constants, which are
@@ -18,11 +20,17 @@ package org.openhab.binding.somfymylink.internal;
  *
  * @author Chris Johnson - Initial contribution
  */
-public class SomfyMyLinkScenesResponse extends SomfyMyLinkResponseBase {
+public class SomfyMyLinkShade {
 
-    public SomfyMyLinkScene[] result;
+    private String targetID;
+    private String name;
 
-    public SomfyMyLinkScene[] getResult() {
-        return result;
+    public String getTargetID() {
+        return targetID.replace('.', '-');
     }
+
+    public String getName() {
+        return name;
+    }
+
 }
